release18.09: add darwin target jobs
These are some parts of the release that I want to get working before
we release 18.09. There have been lots of improvements since 18.03 (as
well as some regressions). To make sure the release is well-tested we
need to add these apps in the jobset. Most of these are UI apps that
are now available.

List of new apps added to the release:

- wireshark
- firefox
- qtmultimedia (already in unstable)
- inkscape (already in unstable)
- gimp
- wireshark
- transmission

Also add ‘stack’. This is one of the Haskell packages hitting the
ARG_MAX limit on macOS (getconf ARG_MAX == 262144). This has not been
solved yet but it will need to be resolved by 18.09. Making it block
here will prevent this regression in the future.

[squashed] release: remove broken from darwin-tested

removes:

- gimp
- qtmultimedia